@@ -39,23 +39,21 @@
     "registry": "https://repository.hybris.com/api/npm/npm-repository/"
   },
   "dependencies": {
-    "@angular/animations": "15.1.1",
-    "@angular/cdk": "15.1.1",
-    "@angular/common": "15.1.1",
-    "@angular/compiler": "15.1.1",
-    "@angular/core": "15.1.1",
-    "@angular/forms": "15.1.1",
-    "@angular/platform-browser": "15.1.1",
-    "@angular/platform-browser-dynamic": "15.1.1",
-    "@angular/router": "15.1.1",
+    "@angular/animations": "15.2.8",
+    "@angular/cdk": "15.2.8",
+    "@angular/common": "15.2.8",
+    "@angular/compiler": "15.2.8",
+    "@angular/core": "15.2.8",
+    "@angular/forms": "15.2.8",
+    "@angular/platform-browser": "15.2.8",
+    "@angular/platform-browser-dynamic": "15.2.8",
+    "@angular/router": "15.2.8",
     "@fundamental-styles/cx": "0.28.3",
     "@fundamental-styles/fn": "0.25.1-rc.16",
-    "@nrwl/angular": "15.5.3",
+    "@nx/angular": "16.1.0",
     "@sap-theming/theming-base-content": "11.3.2",
     "@stackblitz/sdk": "1.8.2",
-    "@swc/helpers": "~0.4.11",
-    "@types/google.visualization": "0.0.68",
-    "@types/node": "16.11.7",
+    "@swc/helpers": "0.5.1",
     "compare-versions": "5.0.1",
     "core-js": "3.26.1",
     "dayjs": "1.11.6",
