@@ -11,11 +11,9 @@ import {
     forwardRef,
     Inject,
     Input,
-    OnDestroy,
-    OnInit,
+    NgZone,
     Optional,
     QueryList,
-    Renderer2,
     SkipSelf,
     ViewChild,
     ViewEncapsulation
@@ -25,19 +23,19 @@ import { DYNAMIC_PAGE_HEADER_TOKEN, DynamicPageHeader } from '@fundamental-ngx/c
 import {
     applyCssClass,
     CssClassBuilder,
+    DestroyedService,
     OVERFLOW_PRIORITY_SCORE,
     OverflowPriority,
-    DestroyedService
+    ResizeObserverService
 } from '@fundamental-ngx/cdk/utils';
-import { fromEvent, Observable, of, Subscription } from 'rxjs';
-import { debounceTime, delay, distinctUntilChanged, filter, takeUntil } from 'rxjs/operators';
-import { ToolbarItemDirective } from './toolbar-item.directive';
+import { BehaviorSubject, combineLatest, map, Observable, startWith, takeUntil } from 'rxjs';
 import { TitleToken } from '@fundamental-ngx/core/title';
 import {
     ContentDensityMode,
     ContentDensityObserver,
     contentDensityObserverProviders
 } from '@fundamental-ngx/core/content-density';
+import { ToolbarItem } from './abstract-toolbar-item.class';
 
 const ELEMENT_MARGIN = 8;
 const OVERFLOW_SPACE = 50 + 2 * ELEMENT_MARGIN;
@@ -66,9 +64,13 @@ export const enum OverflowPriorityEnum {
         DestroyedService
     ]
 })
-export class ToolbarComponent
-    implements OnInit, AfterViewInit, OnDestroy, AfterViewChecked, CssClassBuilder, AfterContentInit
-{
+export class ToolbarComponent implements AfterViewInit, AfterViewChecked, CssClassBuilder, AfterContentInit {
+    /**
+     * The ID of the toolbar title
+     * */
+    @Input()
+    titleId: string;
+
     /** Property allows user to pass additional class
      */
     @Input()
@@ -78,7 +80,14 @@ export class ToolbarComponent
      * Default value: false
      */
     @Input()
-    shouldOverflow = false;
+    set shouldOverflow(val: boolean) {
+        this._shouldOverflow = val;
+        this.shouldOverflow$.next(val);
+    }
+
+    get shouldOverflow(): boolean {
+        return this._shouldOverflow;
+    }
 
     /** Determines the type of toolbar
      * Available options: 'solid' | 'transparent' | 'auto' | 'info'
@@ -86,7 +95,6 @@ export class ToolbarComponent
      */
     @Input()
     fdType: ToolbarType = 'solid';
-
     /** @deprecated */
     @Input()
     hasTitle = false;
@@ -120,26 +128,34 @@ export class ToolbarComponent
     toolbar: ElementRef;
 
     /** @hidden */
-    @ViewChild('overflowBody')
-    overflowBody: ElementRef;
+    @ViewChild('titleElement')
+    titleElement: ElementRef<HTMLHeadElement>;
 
     /** @hidden */
-    @ViewChild('overflowSpacer')
-    overflowSpacer: ElementRef;
-
-    /** @hidden */
-    @ContentChildren(forwardRef(() => ToolbarItemDirective))
-    toolbarItems: QueryList<ToolbarItemDirective>;
+    @ContentChildren(forwardRef(() => ToolbarItem))
+    toolbarItems: QueryList<ToolbarItem>;
 
     /** @hidden */
     @ContentChild(TitleToken)
-    titleComponent: TitleToken | null = null;
+    set titleComponent(title: TitleToken | null) {
+        this._titleComponent$.next(title);
+    }
+
+    get titleComponent(): TitleToken | null {
+        return this._titleComponent$.value;
+    }
 
     /** @hidden */
-    overflowVisibility: Observable<boolean> = of(false);
+    overflowItems$: Observable<ToolbarItem[]>;
 
     /** @hidden */
-    private _subscriptions = new Subscription();
+    overflownItems: ToolbarItem[] = [];
+
+    /** @hidden */
+    private _titleComponent$: BehaviorSubject<TitleToken | null> = new BehaviorSubject<TitleToken | null>(null);
+
+    /** @hidden */
+    private _refreshOverflow$ = new BehaviorSubject<void>(undefined);
 
     /** @hidden */
     private get _toolbarWidth(): number {
@@ -147,65 +163,94 @@ export class ToolbarComponent
     }
 
     /** @hidden */
-    private get _overflowBody(): HTMLElement {
-        return this.overflowBody && (this.overflowBody.nativeElement as HTMLElement);
-    }
-
-    /** @hidden */
-    private get _toolbar(): HTMLElement {
-        return this.toolbar.nativeElement as HTMLElement;
-    }
-
-    /** @hidden */
-    private _overflowElements: ToolbarItemDirective[] = [];
-
-    /** @hidden */
-    private _normalElements: ToolbarItemDirective[] = [];
-
-    /** @hidden */
-    private _disappearElements: ToolbarItemDirective[] = [];
-
-    /** @hidden */
-    private _groupedCollection: { [key: number]: ToolbarItemDirective[] } | null;
-
-    /** @hidden */
-    private _groupedCollectionPriority: { [key: number]: OverflowPriority } | null;
+    private _shouldOverflow = false;
 
     /** @hidden */
     private _initialised = false;
 
     /** @hidden */
+    private shouldOverflow$ = new BehaviorSubject<boolean>(false);
+
+    /** @hidden */
     constructor(
         private _cd: ChangeDetectorRef,
-        private _renderer: Renderer2,
         readonly _contentDensityObserver: ContentDensityObserver,
         private readonly _destroy$: DestroyedService,
+        private resizeObserverService: ResizeObserverService,
+        private ngZone: NgZone,
         @Optional() @SkipSelf() @Inject(DYNAMIC_PAGE_HEADER_TOKEN) private _dynamicPageHeader?: DynamicPageHeader
     ) {
         _contentDensityObserver.subscribe();
     }
 
     /** @hidden */
-    ngOnInit(): void {
-        fromEvent(window, 'resize')
-            .pipe(
-                filter(() => this.shouldOverflow),
-                debounceTime(50),
-                distinctUntilChanged(),
-                takeUntil(this._destroy$)
-            )
-            .subscribe(() => this._onResize());
-    }
-
-    /** @hidden */
     ngAfterViewInit(): void {
-        if (this.shouldOverflow) {
-            of(true)
-                .pipe(delay(5), takeUntil(this._destroy$))
-                .subscribe(() => this._collapseItems());
-        }
+        this.overflowItems$ = combineLatest([
+            this.resizeObserverService.observe(this.toolbar.nativeElement).pipe(map(() => this._toolbarWidth)),
+            this.toolbarItems.changes.pipe(
+                startWith(this.toolbarItems),
+                map((toolbarItems) => toolbarItems.toArray())
+            ),
+            this._titleComponent$.pipe(
+                map((titleComponent): HTMLElement | null => {
+                    if (!titleComponent) {
+                        return this.titleElement?.nativeElement;
+                    }
+                    return titleComponent.elementRef.nativeElement;
+                })
+            ),
+            this.shouldOverflow$,
+            this._refreshOverflow$
+        ]).pipe(
+            map(([toolbarWidth, toolbarItems, titleElement, shouldOverflow]) => {
+                if (shouldOverflow) {
+                    const _sortedByPriorityAndGroupItems = this._getSortedByPriorityAndGroupItems(toolbarItems);
+                    const overflowItems: ToolbarItem[] = [];
+                    _sortedByPriorityAndGroupItems.reduce((_contentWidth, toolbarItem) => {
+                        const itemWidth = toolbarItem.width;
+                        const itemPriority = toolbarItem.priority;
+                        const shouldItemBeRemovedByWidth = itemWidth + _contentWidth > toolbarWidth;
+                        const shouldAlwaysBeInOverflow =
+                            itemPriority === OverflowPriorityEnum.ALWAYS || this.forceOverflow;
+                        const shouldNeverBeInOverflow = itemPriority === OverflowPriorityEnum.NEVER;
+                        if ((shouldItemBeRemovedByWidth && !shouldNeverBeInOverflow) || shouldAlwaysBeInOverflow) {
+                            overflowItems.push(toolbarItem);
+                            return shouldAlwaysBeInOverflow ? _contentWidth : MAX_CONTENT_SIZE;
+                        } else {
+                            return _contentWidth + itemWidth;
+                        }
+                    }, titleElement?.clientWidth || 0);
+                    // Hide orphans
+                    for (const toolbarItem of overflowItems) {
+                        const groupedCollection = this._getGroupedCollection(toolbarItems);
+                        const groupedCollectionPriorities = this._getGroupedCollectionPriority(groupedCollection);
+                        const group = toolbarItem.group;
+                        const itemPriority = toolbarItem.priority;
+                        if (group !== 0 && itemPriority !== OverflowPriorityEnum.DISAPPEAR) {
+                            const itemsFromCurrentGroup =
+                                this._getElementsFromCurrentGroup(
+                                    group,
+                                    groupedCollection,
+                                    groupedCollectionPriorities
+                                ) || [];
+                            itemsFromCurrentGroup.forEach((i) => {
+                                if (overflowItems.indexOf(i) === -1) {
+                                    overflowItems.push(i);
+                                }
+                            });
+                        }
+                    }
+                    return overflowItems.sort((a, b) => this._sortPriorities(a.priority, b.priority));
+                }
+                return [];
+            }),
+            takeUntil(this._destroy$)
+        );
+        this.overflowItems$.subscribe((items) => {
+            this.overflownItems = items;
+            this._cd.detectChanges();
+        });
         this._initialised = true;
-        this._listenForItemChanges();
         this.buildComponentCssClass();
     }
 
@@ -220,11 +265,6 @@ export class ToolbarComponent
             this.titleComponent.elementRef?.nativeElement.classList.add('fd-toolbar__title');
             this.buildComponentCssClass();
         }
-    }
-
-    /** @hidden */
-    ngOnDestroy(): void {
-        this._subscriptions.unsubscribe();
     }
 
     /** Change detection reference */
@@ -252,141 +292,110 @@ export class ToolbarComponent
 
     /** Method triggering collapsible toolbar  */
     updateCollapsibleItems(): void {
-        if (this._initialised) {
-            this._onResize();
-        }
+        this._refreshOverflow$.next();
     }
 
-    /** @hidden */
-    private _onResize(): void {
-        this._reset();
-        if (this.shouldOverflow) {
-            this._collapseItems();
+    /** @hidden Get group number with the lowest priority.
+     *  Uses for detecting a group of elements which would be hidden from the toolbar.
+     *  */
+    private _getLowestPriorityFromGroup(
+        itemGroup: number,
+        groupedCollectionPriority?: Record<number, OverflowPriority>
+    ): number {
+        const priority = groupedCollectionPriority;
+        if (!priority || !OVERFLOW_PRIORITY_SCORE.get(priority[itemGroup])) {
+            return itemGroup;
         }
-        this._cd.detectChanges();
-    }
 
-    // shouldOverflow items
-    /** @hidden */
-    private _collapseItems(): void {
-        this._groupedCollection = this._getGroupedCollection();
-        this._groupedCollectionPriority = this._getGroupedCollectionPriority();
-        const _sortedByPriorityAndGroupItems = this._getSortedByPriorityAndGroupItems();
+        const groups = Object.keys(priority);
 
-        _sortedByPriorityAndGroupItems.reduce((_contentWidth, toolbarItem) => {
-            const itemWidth = this._getElementWidthWithMargin(toolbarItem);
-            const itemGroup = this._getElementGroup(toolbarItem);
-            const itemPriority = this._getElementPriority(toolbarItem);
-            const shouldItemBeRemovedByWidth = this._shouldToolbarItemBeRemovedByWidth(itemWidth, _contentWidth);
-            const shouldAlwaysBeInOverflow = itemPriority === OverflowPriorityEnum.ALWAYS || this.forceOverflow;
-            const shouldNeverBeInOverflow = itemPriority === OverflowPriorityEnum.NEVER;
-            if ((shouldItemBeRemovedByWidth && !shouldNeverBeInOverflow) || shouldAlwaysBeInOverflow) {
-                if (itemGroup) {
-                    this._hideElementsFromCurrentGroup(itemGroup);
-                } else {
-                    this._hideElementWithoutGroup(toolbarItem, itemPriority);
-                }
-                return shouldAlwaysBeInOverflow ? _contentWidth : MAX_CONTENT_SIZE;
-            } else {
-                this._normalElements.push(toolbarItem);
-                return _contentWidth + itemWidth;
+        return groups.reduce((lowestGroup, currentGroup) => {
+            const lowestGroupPriority = priority[lowestGroup];
+            const currentGroupPriority = priority[currentGroup];
+            if (
+                OVERFLOW_PRIORITY_SCORE.get(currentGroupPriority)! < OVERFLOW_PRIORITY_SCORE.get(lowestGroupPriority)!
+            ) {
+                return +currentGroup;
             }
-        }, 0);
 
-        this._collapseLastSpacerElement(this._overflowElements, this._normalElements);
-        this._addToolbarItemToOverflow(this._overflowElements);
-        this._disappearElements.forEach(this._removeToolbarItemFromDOM.bind(this));
-
-        this._normalElements = [...this._getOrderedItemsFollowingQuery(this._normalElements)];
-        this._overflowElements = [...this._getOrderedItemsFollowingQuery(this._overflowElements)];
-
-        this.toolbarItems.forEach((item) => this._changeItemVisibilityState(item.elementRef.nativeElement, true));
-
-        this._changeOverflowVisibleState(this._overflowElements.length > 0);
-
-        this._cd.markForCheck();
+            return lowestGroup;
+        }, itemGroup as number);
     }
 
     /** @hidden */
-    private _shouldToolbarItemBeRemovedByWidth(itemWidth: number, contentWidth: number): boolean {
-        return contentWidth + itemWidth > this._toolbarWidth;
-    }
-
-    /** @hidden */
-    private _getElementWidthWithMargin(toolbarItem: ToolbarItemDirective): number {
-        return (
-            toolbarItem.elementRef.nativeElement && toolbarItem.elementRef.nativeElement.offsetWidth + ELEMENT_MARGIN
-        );
-    }
-
-    /** @hidden */
-    private _listenForItemChanges(): void {
-        this.toolbarItems.changes.pipe(filter(() => this.shouldOverflow)).subscribe(() => this._onResize());
-    }
-
-    /** @hidden */
-    private _getElementPriority(toolbarItem: ToolbarItemDirective): OverflowPriority {
-        const priority =
-            toolbarItem.elementRef.nativeElement && toolbarItem.elementRef.nativeElement.attributes.fdOverflowPriority;
-
-        return (priority && priority.value) || OverflowPriorityEnum.HIGH;
-    }
-
-    /** @hidden */
-    private _getElementGroup(toolbarItem: ToolbarItemDirective): number {
-        const itemGroup =
-            toolbarItem.elementRef.nativeElement && toolbarItem.elementRef.nativeElement.attributes.fdOverflowGroup;
-
-        return (itemGroup && itemGroup.value) || 0;
-    }
-
-    /** @hidden */
-    private _hideElementsFromCurrentGroup(itemGroup: number): void {
-        const lowestPriorityFromGroup = this._getLowestPriorityFromGroup(itemGroup);
-        if (!this._groupedCollection || !this._groupedCollectionPriority) {
-            return;
+    private _getElementsFromCurrentGroup(
+        itemGroup: number,
+        groupedCollection?: Record<number, ToolbarItem[]>,
+        groupedCollectionPriority?: Record<number, OverflowPriority>
+    ): ToolbarItem[] {
+        const lowestPriorityFromGroup = this._getLowestPriorityFromGroup(itemGroup, groupedCollectionPriority);
+        if (!groupedCollection || !groupedCollectionPriority) {
+            return [];
         }
-        let itemsToRemove = this._groupedCollection[lowestPriorityFromGroup];
+        let itemsToRemove = groupedCollection[lowestPriorityFromGroup];
 
-        delete this._groupedCollection[lowestPriorityFromGroup];
-        delete this._groupedCollectionPriority[lowestPriorityFromGroup];
+        delete groupedCollection[lowestPriorityFromGroup];
+        delete groupedCollectionPriority[lowestPriorityFromGroup];
 
         if (lowestPriorityFromGroup !== itemGroup) {
-            itemsToRemove = [...itemsToRemove, ...this._groupedCollection[itemGroup]];
-            delete this._groupedCollection[itemGroup];
-            delete this._groupedCollectionPriority[itemGroup];
+            itemsToRemove = [...itemsToRemove, ...groupedCollection[itemGroup]];
+            delete groupedCollection[itemGroup];
+            delete groupedCollectionPriority[itemGroup];
         }
 
-        itemsToRemove?.forEach((item) => {
-            this._removeItemFromArray(this._normalElements, item);
-            const isDisappear = this._getElementPriority(item) === OverflowPriorityEnum.DISAPPEAR;
-            if (isDisappear) {
-                this._disappearElements.push(item);
-            }
-
-            if (this._overflowElements.indexOf(item) === -1 && !isDisappear) {
-                this._overflowElements.push(item);
-            }
-        });
+        return itemsToRemove;
     }
 
-    /** @hidden */
-    private _hideElementWithoutGroup(toolbarItem: ToolbarItemDirective, itemPriority: OverflowPriority): void {
-        const isDisappear = itemPriority === OverflowPriorityEnum.DISAPPEAR;
-        if (isDisappear) {
-            this._disappearElements.push(toolbarItem);
-        } else {
-            this._overflowElements.push(toolbarItem);
-        }
+    /** @hidden Get the object with grouped arrays of elements. */
+    private _getGroupedCollection(toolbarItems: ToolbarItem[]): Record<number, ToolbarItem[]> {
+        return toolbarItems.reduce((acc, item) => {
+            const itemPrio = item.priority;
+            const itemGroup = item.group;
+            if (
+                itemPrio === OverflowPriorityEnum.NEVER ||
+                itemPrio === OverflowPriorityEnum.ALWAYS ||
+                itemGroup === 0
+            ) {
+                return acc;
+            }
+            if (!acc[itemGroup]) {
+                acc[itemGroup] = [];
+            }
+            acc[itemGroup].push(item);
+            return acc;
+        }, {});
+    }
+
+    /** @hidden Get the object with the highest priority for each group of elements. */
+    private _getGroupedCollectionPriority(
+        groupedCollection: Record<number, ToolbarItem[]>
+    ): Record<number, OverflowPriority> {
+        const groups = Object.keys(groupedCollection);
+
+        return groups.reduce((acc, itemGroup) => {
+            const sortedPriorities = groupedCollection![itemGroup]
+                .map((i) => i.priority)
+                .filter((prio) => prio !== OverflowPriorityEnum.ALWAYS && prio !== OverflowPriorityEnum.NEVER)
+                .sort(this._sortPriorities);
+
+            acc[itemGroup] = sortedPriorities[0];
+
+            return acc;
+        }, {});
+    }
+
+    /** @hidden Sort priorities of elements/groups. */
+    private _sortPriorities(a: OverflowPriority, b: OverflowPriority): number {
+        return OVERFLOW_PRIORITY_SCORE.get(b)! - OVERFLOW_PRIORITY_SCORE.get(a)!;
     }
 
     /** @hidden Sort by group and priority and initial position */
-    private _getSortedByPriorityAndGroupItems(): ToolbarItemDirective[] {
-        const notSorted = this.toolbarItems.toArray().map((element, index) => ({ element, index }));
+    private _getSortedByPriorityAndGroupItems(toolbarItems: ToolbarItem[]): ToolbarItem[] {
+        const notSorted = toolbarItems.map((element, index) => ({ element, index }));
 
         const groups = notSorted.reduce((gr, item) => {
-            let groupId = this._getElementGroup(item.element);
-            const itemPrio = this._getElementPriority(item.element);
+            let groupId = item.element.group;
+            const itemPrio = item.element.priority;
             if (itemPrio === OverflowPriorityEnum.NEVER || itemPrio === OverflowPriorityEnum.ALWAYS) {
                 groupId = 0;
             }
@@ -412,7 +421,7 @@ export class ToolbarComponent
                     minIndex = Math.min(minIndex, item.index);
                     maxPriority = Math.max(
                         maxPriority,
-                        OVERFLOW_PRIORITY_SCORE.get(this._getElementPriority(item.element)) ?? -Infinity
+                        OVERFLOW_PRIORITY_SCORE.get(item.element.priority) ?? -Infinity
                     );
                 }
 
@@ -423,181 +432,11 @@ export class ToolbarComponent
                     ? []
                     : groups[0].map((item) => ({
                           group: [item.element],
-                          maxPriority: OVERFLOW_PRIORITY_SCORE.get(this._getElementPriority(item.element)),
+                          maxPriority: OVERFLOW_PRIORITY_SCORE.get(item.element.priority),
                           minIndex: item.index
                       }))
             )
             .sort((a, b) => b.maxPriority - a.maxPriority || a.minIndex - b.minIndex)
             .reduce((arr, i) => arr.concat(i.group), []);
-    }
-
-    /** @hidden Get the object with grouped arrays of elements. */
-    private _getGroupedCollection(): { [key: number]: ToolbarItemDirective[] } {
-        const collection =
-            (this.toolbarItems &&
-                this.toolbarItems.reduce((acc, item) => {
-                    const itemPrio = this._getElementPriority(item);
-                    if (itemPrio === OverflowPriorityEnum.NEVER || itemPrio === OverflowPriorityEnum.ALWAYS) {
-                        return acc;
-                    }
-
-                    const itemGroup = this._getElementGroup(item);
-                    acc[itemGroup] && acc[itemGroup].length ? acc[itemGroup].push(item) : (acc[itemGroup] = [item]);
-
-                    return acc;
-                }, {})) ||
-            {};
-
-        delete collection[0];
-
-        return collection;
-    }
-
-    /** @hidden Get the object with the highest priority for each group of elements. */
-    private _getGroupedCollectionPriority(): { [key: number]: OverflowPriority } {
-        if (!this._groupedCollection) {
-            return {};
-        }
-
-        const groups = Object.keys(this._groupedCollection);
-
-        return groups.reduce((acc, itemGroup) => {
-            const sortedPriorities = this._groupedCollection![itemGroup].map(this._getElementPriority)
-                .filter((prio) => prio !== OverflowPriorityEnum.ALWAYS && prio !== OverflowPriorityEnum.NEVER)
-                .sort(this._sortPriorities.bind(this));
-
-            acc[itemGroup] = sortedPriorities[0];
-
-            return acc;
-        }, {});
-    }
-
-    /** @hidden Get group number with the lowest priority.
-     *  Uses for detecting a group of elements which would be hidden from the toolbar.
-     *  */
-    private _getLowestPriorityFromGroup(itemGroup: number): number {
-        const priority = this._groupedCollectionPriority;
-        if (!priority || !OVERFLOW_PRIORITY_SCORE.get(priority[itemGroup])) {
-            return itemGroup;
-        }
-
-        const groups = Object.keys(priority);
-
-        return groups.reduce((lowestGroup, currentGroup) => {
-            const lowestGroupPriority = priority[lowestGroup];
-            const currentGroupPriority = priority[currentGroup];
-            if (
-                OVERFLOW_PRIORITY_SCORE.get(currentGroupPriority)! < OVERFLOW_PRIORITY_SCORE.get(lowestGroupPriority)!
-            ) {
-                return +currentGroup;
-            }
-
-            return lowestGroup;
-        }, itemGroup as number);
-    }
-
-    /** @hidden */
-    private _removeItemFromArray(
-        toolbarItems: ToolbarItemDirective[],
-        toolbarItem: ToolbarItemDirective
-    ): ToolbarItemDirective | null {
-        const itemIndex = toolbarItems.indexOf(toolbarItem);
-
-        return itemIndex >= 0 ? toolbarItems.splice(itemIndex, 1)[0] : null;
-    }
-
-    /** @hidden */
-    private _removeToolbarItemFromDOM(toolbarItem: ToolbarItemDirective): void {
-        if (toolbarItem.elementRef.nativeElement && toolbarItem.elementRef.nativeElement.parentNode) {
-            // IE11 workaround element.remove() is not supported
-            this._renderer.removeChild(this.elementRef().nativeElement, toolbarItem.elementRef.nativeElement);
-        }
-    }
-
-    /** @hidden */
-    private _addToolbarItemToOverflow(toolbarItems: ToolbarItemDirective[]): void {
-        toolbarItems.forEach((x) => {
-            if (this._overflowBody) {
-                this._overflowBody.appendChild(x.elementRef.nativeElement);
-            }
-        });
-    }
-
-    /** @hidden */
-    private _reset(): void {
-        this._normalElements.forEach(this._removeToolbarItemFromDOM.bind(this));
-        this._overflowElements.forEach(this._removeToolbarItemFromDOM.bind(this));
-
-        if (this.overflowSpacer) {
-            this.toolbarItems.map((x) => {
-                this._renderer.insertBefore(
-                    this._toolbar,
-                    x.elementRef.nativeElement,
-                    this.overflowSpacer.nativeElement
-                );
-            });
-        }
-
-        this._overflowElements = [];
-        this._normalElements = [];
-        this._disappearElements = [];
-
-        this._groupedCollection = null;
-        this._groupedCollectionPriority = null;
-
-        this._changeOverflowVisibleState(false);
-    }
-
-    /** @hidden */
-    private _changeOverflowVisibleState(visible: boolean): void {
-        this.overflowVisibility = of(visible).pipe(debounceTime(1));
-    }
-
-    /** @hidden */
-    private _changeItemVisibilityState(element: HTMLElement, visible: boolean): void {
-        const fadeIn = 'fd-toolbar-fade-in';
-        const fadeOut = 'fd-toolbar-fade-out';
-
-        if (visible) {
-            element.classList.add(fadeIn);
-            element.classList.remove(fadeOut);
-            return;
-        }
-
-        element.classList.add(fadeOut);
-        element.classList.remove(fadeIn);
-    }
-
-    /** @hidden Sort priorities of elements/groups. */
-    private _sortPriorities(a: OverflowPriority, b: OverflowPriority): number {
-        return OVERFLOW_PRIORITY_SCORE.get(b)! - OVERFLOW_PRIORITY_SCORE.get(a)!;
-    }
-
-    /** @hidden */
-    private _collapseLastSpacerElement(
-        overflowElements: ToolbarItemDirective[],
-        normalElements: ToolbarItemDirective[]
-    ): void {
-        if (!normalElements || normalElements.length < 1) {
-            return;
-        }
-
-        const isSpacer = (element: ToolbarItemDirective): boolean =>
-            element.elementRef.nativeElement.classList.contains('fd-toolbar__spacer');
-
-        let lastItem: ToolbarItemDirective = normalElements[normalElements.length - 1];
-
-        while (isSpacer(lastItem) && normalElements.length > 0) {
-            overflowElements.push(normalElements.pop()!);
-            lastItem = normalElements[normalElements.length - 1];
-        }
-    }
-
-    /** @hidden */
-    private _getOrderedItemsFollowingQuery(toolbarItems: ToolbarItemDirective[]): ToolbarItemDirective[] {
-        const queryArray = this.toolbarItems.toArray();
-        const sortMethod = (a: ToolbarItemDirective, b: ToolbarItemDirective): number =>
-            queryArray.findIndex((_item) => _item === a) > queryArray.findIndex((_item) => _item === b) ? 1 : -1;
-        return toolbarItems.sort(sortMethod);
     }
 }
