@@ -1,6 +1,5 @@
 const { getInput, info } = require('@actions/core');
 const { npmPublish } = require('@jsdevtools/npm-publish');
-const { readFileSync } = require('fs');
 const { resolve } = require('path');
 
 async function publish({ currentTryNumber = 1, packageJsonPath, tag, token, access, retryCount }) {
@@ -28,14 +27,8 @@ async function publish({ currentTryNumber = 1, packageJsonPath, tag, token, acce
 }
 
 const run = async () => {
-    const projectsMap = JSON.parse(readFileSync('./angular.json', { encoding: 'utf-8' })).projects;
     const projectNames = JSON.parse(getInput('projects'));
-    const projects = projectNames.map((projectName) => {
-        const ngPackage = JSON.parse(
-            readFileSync(`${projectsMap[projectName]}/ng-package.json`, { encoding: 'utf-8' })
-        );
-        return resolve(projectsMap[projectName], ngPackage.dest, 'package.json');
-    });
+    const projects = projectNames.map((projectName) => resolve(`dist/libs/${projectName}/package.json`));
     const tag = getInput('releaseTag');
     const npmToken = getInput('token');
     const retryCount = 3;
