@@ -27,10 +27,10 @@ const cp = require('child_process');
 const isWindows = os.platform() === 'win32';
 let output;
 try {
-    output = require('@nrwl/workspace').output;
+    output = require('@nx/workspace').output;
 } catch (e) {
     console.warn(
-        'Angular CLI could not be decorated to enable computation caching. Please ensure @nrwl/workspace is installed.'
+        'Angular CLI could not be decorated to enable computation caching. Please ensure @nx/workspace is installed.'
     );
     process.exit(0);
 }
