@@ -1,4 +1,4 @@
-import { Directive, HostBinding } from '@angular/core';
+import { Directive, ElementRef, HostBinding } from '@angular/core';
 
 @Directive({
     selector: '[fd-list-navigation-item-arrow], [fdListNavigaitonItemArrow]'
@@ -21,11 +21,19 @@ export class ListNavigationItemArrowDirective {
     expanded = false;
 
     /** @hidden */
+    constructor(private _elRef: ElementRef) {}
+
+    /** @hidden */
     _setExpanded(expanded: boolean): void {
         if (this.expanded !== expanded) {
             this.rightArrowClass = !this.rightArrowClass;
             this.downArrowClass = !this.downArrowClass;
         }
         this.expanded = expanded;
+    }
+
+    /** @hidden */
+    _focus(): void {
+        this._elRef.nativeElement.focus();
     }
 }
