@@ -1,13 +1,12 @@
 export * from './toolbar.module';
 export * from './toolbar.component';
 export * from './toolbar-item.directive';
-export * from './toolbar-spacer.component';
+export * from './toolbar-spacer.directive';
 export * from './toolbar-separator.component';
 export * from './toolbar-label.directive';
 export * from './toolbar-form-label.directive';
-export * from './toolbar-overflow-priority.directive';
 export * from './toolbar-form-label.directive';
-export * from './toolbar-overflow-group.directive';
 export * from './toolbar-overflow-button.directive';
 export * from './toolbar-overflow-button-menu.directive';
 export * from './deprecated-toolbar-size.directive';
+export * from './abstract-toolbar-item.class';
