@@ -1,8 +1,10 @@
 import { Component, QueryList, ViewChild, ViewChildren } from '@angular/core';
 import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
+import { firstValueFrom } from 'rxjs';
 
-import { ButtonModule } from '@fundamental-ngx/core/button';
 import { whenStable } from '@fundamental-ngx/core/tests';
+import { OverflowPriority } from '@fundamental-ngx/cdk/utils';
+
 import { ToolbarItemDirective } from './toolbar-item.directive';
 import { ToolbarComponent } from './toolbar.component';
 import { ToolbarModule } from './toolbar.module';
@@ -15,7 +17,7 @@ describe('ToolbarComponent', () => {
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
             declarations: [ToolbarTestComponent],
-            imports: [ToolbarModule, ButtonModule]
+            imports: [ToolbarModule]
         }).compileComponents();
     }));
 
@@ -26,7 +28,6 @@ describe('ToolbarComponent', () => {
 
         component = fixture.componentInstance;
         toolbar = component.toolbar;
-        toolbar.toolbarItems = component.childrens;
 
         await whenStable(fixture);
     });
@@ -36,7 +37,9 @@ describe('ToolbarComponent', () => {
     });
 
     it('should overflow', async () => {
-        expect(toolbar['_overflowElements'].length).toBeGreaterThan(0);
+        await fixture.whenRenderingDone();
+        const actualOverflowItems = await firstValueFrom(toolbar.overflowItems$);
+        expect(actualOverflowItems.length).toBeGreaterThan(0);
     });
 });
 
@@ -47,7 +50,7 @@ describe('ToolbarComponent - Prioritization', () => {
 
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
-            imports: [ToolbarModule, ButtonModule],
+            imports: [ToolbarModule],
             declarations: [ToolbarOverflowPriorityTestComponent]
         }).compileComponents();
     }));
@@ -59,29 +62,18 @@ describe('ToolbarComponent - Prioritization', () => {
 
         component = fixture.componentInstance;
         toolbar = component.toolbar;
-        toolbar.toolbarItems = component.childrens;
+        toolbar.toolbarItems = component.children;
 
         await whenStable(fixture);
     });
 
     it('should hide element to overflow by priority', async () => {
-        const normalElements = ['Button First', 'Never', 'High'];
-        const overflowElements = ['Always', 'Low', 'Button Last'];
-        const disappearElements = ['Disappear'];
+        await fixture.whenRenderingDone();
+        const overflownItems: OverflowPriority[] = ['high', 'low', 'disappear', 'always'];
+        const actualOverflownItems = await firstValueFrom(toolbar.overflowItems$);
+        expect(actualOverflownItems.length).toBeGreaterThan(0);
 
-        expect(toolbar['_overflowElements'].length).toBeGreaterThan(0);
-        expect(toolbar['_normalElements'].length).toBeGreaterThan(0);
-        expect(toolbar['_disappearElements'].length).toBeGreaterThan(0);
-
-        expect(toolbar['_overflowElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            overflowElements
-        );
-        expect(toolbar['_normalElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            normalElements
-        );
-        expect(toolbar['_disappearElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            disappearElements
-        );
+        expect(actualOverflownItems.map((el) => el.priority)).toEqual(overflownItems);
     });
 });
 
@@ -92,7 +84,7 @@ describe('ToolbarComponent - Prioritization and Grouping', () => {
 
     beforeEach(waitForAsync(() => {
         TestBed.configureTestingModule({
-            imports: [ToolbarModule, ButtonModule],
+            imports: [ToolbarModule],
             declarations: [ToolbarOverflowGroupingTestComponent]
         }).compileComponents();
     }));
@@ -104,29 +96,21 @@ describe('ToolbarComponent - Prioritization and Grouping', () => {
 
         component = fixture.componentInstance;
         toolbar = component.toolbar;
-        toolbar.toolbarItems = component.childrens;
 
         await whenStable(fixture);
     });
 
     it('should hide elements to overflow by group and priority', async () => {
-        const normalElements = ['Button', 'Never', 'Gr 1 / Low', 'Gr 1 / High'];
-        const overflowElements = ['Always', 'Gr 2 / Low', 'Gr 2 / Low'];
-        const disappearElements = ['Gr 2 / Disappear'];
+        const expectedOverflownItems: `${number} / ${OverflowPriority}`[] = [
+            '2 / low',
+            '2 / low',
+            '2 / disappear',
+            '0 / always'
+        ];
+        const actualOverflownItems = await firstValueFrom(toolbar.overflowItems$);
+        expect(actualOverflownItems.length).toBeGreaterThan(0);
 
-        expect(toolbar['_overflowElements'].length).toBeGreaterThan(0);
-        expect(toolbar['_normalElements'].length).toBeGreaterThan(0);
-        expect(toolbar['_disappearElements'].length).toBeGreaterThan(0);
-
-        expect(toolbar['_overflowElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            overflowElements
-        );
-        expect(toolbar['_normalElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            normalElements
-        );
-        expect(toolbar['_disappearElements'].map((el) => el.elementRef.nativeElement.textContent?.trim())).toEqual(
-            disappearElements
-        );
+        expect(actualOverflownItems.map((el) => `${el.group} / ${el.priority}`)).toEqual(expectedOverflownItems);
     });
 });
 
@@ -135,35 +119,35 @@ describe('ToolbarComponent - Prioritization and Grouping', () => {
     template: `
         <div [style.width]="width">
             <fd-toolbar #toolbar [shouldOverflow]="true">
-                <button fd-toolbar-item fd-button fdCompact>Button1</button>
+                <button fd-toolbar-item>Button1</button>
                 <fd-toolbar-separator fd-toolbar-item></fd-toolbar-separator>
-                <button fd-toolbar-item fd-button fdCompact>Button2</button>
+                <button fd-toolbar-item>Button2</button>
                 <fd-toolbar-separator fd-toolbar-item></fd-toolbar-separator>
-                <button fd-toolbar-item fd-button fdCompact>Button3</button>
-                <button fd-toolbar-item fd-button fdCompact>Button4</button>
+                <button fd-toolbar-item>Button3</button>
+                <button fd-toolbar-item>Button4</button>
 
                 <fd-toolbar-spacer fd-toolbar-item></fd-toolbar-spacer>
 
                 <fd-toolbar-separator fd-toolbar-item></fd-toolbar-separator>
 
-                <button fd-toolbar-item fd-button fdCompact>Button5</button>
-                <button fd-toolbar-item fd-button fdCompact>Button6</button>
-                <button fd-toolbar-item fd-button fdCompact>Button7</button>
-                <button fd-toolbar-item fd-button fdCompact>Button8</button>
-                <button fd-toolbar-item fd-button fdCompact>Button9</button>
-                <button fd-toolbar-item fd-button fdCompact>Button10</button>
+                <button fd-toolbar-item>Button5</button>
+                <button fd-toolbar-item>Button6</button>
+                <button fd-toolbar-item>Button7</button>
+                <button fd-toolbar-item>Button8</button>
+                <button fd-toolbar-item>Button9</button>
+                <button fd-toolbar-item>Button10</button>
                 <fd-toolbar-spacer fd-toolbar-item></fd-toolbar-spacer>
-                <button fd-toolbar-item fd-button fdCompact>Button11</button>
-                <button fd-toolbar-item fd-button fdCompact>Button12</button>
+                <button fd-toolbar-item>Button11</button>
+                <button fd-toolbar-item>Button12</button>
                 <fd-toolbar-separator fd-toolbar-item></fd-toolbar-separator>
-                <button fd-toolbar-item fd-button fdCompact>Button13</button>
-                <button fd-toolbar-item fd-button fdCompact>Button14</button>
-                <button fd-toolbar-item fd-button fdCompact>Button15</button>
-                <button fd-toolbar-item fd-button fdCompact>Button16</button>
-                <button fd-toolbar-item fd-button fdCompact>Button17</button>
-                <button fd-toolbar-item fd-button fdCompact>Button18</button>
-                <button fd-toolbar-item fd-button fdCompact>Button19</button>
-                <button fd-toolbar-item fd-button fdCompact>Button20</button>
+                <button fd-toolbar-item>Button13</button>
+                <button fd-toolbar-item>Button14</button>
+                <button fd-toolbar-item>Button15</button>
+                <button fd-toolbar-item>Button16</button>
+                <button fd-toolbar-item>Button17</button>
+                <button fd-toolbar-item>Button18</button>
+                <button fd-toolbar-item>Button19</button>
+                <button fd-toolbar-item>Button20</button>
                 <fd-toolbar-separator fd-toolbar-item></fd-toolbar-separator>
             </fd-toolbar>
         </div>
@@ -171,7 +155,7 @@ describe('ToolbarComponent - Prioritization and Grouping', () => {
 })
 class ToolbarTestComponent {
     @ViewChild('toolbar') toolbar: ToolbarComponent;
-    @ViewChildren(ToolbarItemDirective) childrens: QueryList<ToolbarItemDirective>;
+    @ViewChildren(ToolbarItemDirective) children: QueryList<ToolbarItemDirective>;
 
     width = '300px';
 }
@@ -181,20 +165,20 @@ class ToolbarTestComponent {
     template: `
         <div [style.width]="width">
             <fd-toolbar #toolbar [shouldOverflow]="true">
-                <button fd-toolbar-item fd-button fdCompact>Button First</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="always">Always</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="never">Never</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="low">Low</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="high">High</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="disappear">Disappear</button>
-                <button fd-toolbar-item fd-button fdCompact>Button Last</button>
+                <button fd-toolbar-item>Button First</button>
+                <button fd-toolbar-item fdOverflowPriority="always">Always</button>
+                <button fd-toolbar-item fdOverflowPriority="never">Never</button>
+                <button fd-toolbar-item fdOverflowPriority="low">Low</button>
+                <button fd-toolbar-item fdOverflowPriority="high">High</button>
+                <button fd-toolbar-item fdOverflowPriority="disappear">Disappear</button>
+                <button fd-toolbar-item>Button Last</button>
             </fd-toolbar>
         </div>
     `
 })
 class ToolbarOverflowPriorityTestComponent {
     @ViewChild('toolbar') toolbar: ToolbarComponent;
-    @ViewChildren(ToolbarItemDirective) childrens: QueryList<ToolbarItemDirective>;
+    @ViewChildren(ToolbarItemDirective) children: QueryList<ToolbarItemDirective>;
 
     width = '300px';
 }
@@ -204,31 +188,21 @@ class ToolbarOverflowPriorityTestComponent {
     template: `
         <div [style.width]="width">
             <fd-toolbar #toolbar [shouldOverflow]="true">
-                <button fd-toolbar-item fd-button fdCompact>Button</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="always">Always</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="never">Never</button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="low" fdOverflowGroup="1">
-                    Gr 1 / Low
-                </button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="low" fdOverflowGroup="2">
-                    Gr 2 / Low
-                </button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="disappear" fdOverflowGroup="2">
-                    Gr 2 / Disappear
-                </button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="low" fdOverflowGroup="2">
-                    Gr 2 / Low
-                </button>
-                <button fd-toolbar-item fd-button fdCompact fdOverflowPriority="high" fdOverflowGroup="1">
-                    Gr 1 / High
-                </button>
+                <button fd-toolbar-item>Button</button>
+                <button fd-toolbar-item fdOverflowPriority="always">Always</button>
+                <button fd-toolbar-item fdOverflowPriority="never">Never</button>
+                <button fd-toolbar-item fdOverflowPriority="low" [fdOverflowGroup]="1">Gr 1 / Low</button>
+                <button fd-toolbar-item fdOverflowPriority="low" [fdOverflowGroup]="2">Gr 2 / Low</button>
+                <button fd-toolbar-item fdOverflowPriority="disappear" [fdOverflowGroup]="2">Gr 2 / Disappear</button>
+                <button fd-toolbar-item fdOverflowPriority="low" [fdOverflowGroup]="2">Gr 2 / Low</button>
+                <button fd-toolbar-item fdOverflowPriority="high" [fdOverflowGroup]="1">Gr 1 / High</button>
             </fd-toolbar>
         </div>
     `
 })
 class ToolbarOverflowGroupingTestComponent {
     @ViewChild('toolbar') toolbar: ToolbarComponent;
-    @ViewChildren(ToolbarItemDirective) childrens: QueryList<ToolbarItemDirective>;
+    @ViewChildren(ToolbarItemDirective) children: QueryList<ToolbarItemDirective>;
 
     width = '400px';
 }
