@@ -1,5 +1,5 @@
 import { AvatarGroupPo } from './avatar-group.po';
-// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+// eslint-disable-next-line @nx/enforce-module-boundaries
 import {
     checkElArrIsClickable,
     click,
