@@ -1,5 +1,5 @@
 import { ComboboxPo } from './combobox.po';
-// eslint-disable-next-line @nrwl/nx/enforce-module-boundaries
+// eslint-disable-next-line @nx/enforce-module-boundaries
 import {
     acceptAlert,
     browserIsSafari,
