@@ -2,14 +2,16 @@
 export default {
     displayName: 'nx-plugin',
     preset: '../../jest.preset.js',
-    globals: {
-        'ts-jest': {
-            tsconfig: '<rootDir>/tsconfig.spec.json'
-        }
-    },
+    globals: {},
     transform: {
-        '^.+\\.[tj]s$': 'ts-jest'
+        '^.+\\.[tj]s$': [
+            'ts-jest',
+            {
+                tsconfig: '<rootDir>/tsconfig.spec.json'
+            }
+        ]
     },
+    transformIgnorePatterns: ['node_modules/(?!.*\\.mjs$)'],
     moduleFileExtensions: ['ts', 'js', 'html'],
     coverageDirectory: '../../coverage/libs/nx-plugin'
 };
