@@ -563,8 +563,13 @@ export abstract class BaseMultiCombobox extends CollectionBaseInput implements O
         } else if (!KeyUtil.isKeyCode(event, [...this._nonOpeningKeys, CONTROL])) {
             this.showList(true);
             const acceptedKeys = !KeyUtil.isKeyType(event, 'alphabetical') && !KeyUtil.isKeyType(event, 'numeric');
-            if (this.isEmptyValue && acceptedKeys) {
-                this.listComponent?.setItemActive(0);
+            if (acceptedKeys) {
+                // SetTimeout is needed for input to receive new value.
+                setTimeout(() => {
+                    if (this.isEmptyValue) {
+                        this.listComponent?.setItemActive(0);
+                    }
+                });
             }
         }
     }
