@@ -1,5 +1,5 @@
 export * from './panel.component';
 export * from './panel-content/panel-content.component';
-export * from './panel-actions/panel-actions.component';
+export * from './panel-actions.component';
 export * from './panel.config';
 export * from './panel.module';
